Split core functionality into a common service
