+++ apps/ues-angular/browserslist
@@ -5,10 +5,9 @@
 # You can see what browsers were selected by your queries by running:
 #   npx browserslist
 
-> 0.5%
 last 2 versions
+ios_saf >= 10
+Safari >= 10
 Firefox ESR
 not dead
 not IE 9-11 # For IE 9-11 support, remove 'not'.
-ios_saf >= 10
-Safari >= 10
