+++ libs/ues/common/ngx/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/ues-common-ngx.module';
+export * from './lib/types/types';
 export * from './lib/modules/routing/routing.module';
 export * from './lib/modules/sidebar/sidebar.module';
 export * from './lib/modules/core-ui/core-ui.module';
