+++ libs/ues/effluent/src/lib/modules/core/core.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@NgModule({
+  imports: [CommonModule],
+  declarations: []
+})
+export class CoreModule {}
