+++ libs/ues/effluent/src/lib/modules/core/services/effluent.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { EffluentService } from './effluent.service';
+
+describe('EffluentService', () => {
+  let service: EffluentService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(EffluentService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
